Cover BillableTrait in tests
