+++ tests/TestCase.php
@@ -2,9 +2,15 @@
 
 namespace Payavel\Checkout\Tests;
 
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
+use Illuminate\Support\Facades\Schema;
+use Payavel\Checkout\Contracts\Billable;
 use Payavel\Checkout\PaymentServiceProvider;
+use Payavel\Checkout\Traits\Billable as BillableTrait;
 
 abstract class TestCase extends \Orchestra\Testbench\TestCase
 {
@@ -38,5 +44,46 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
 
             $this->artisan('migrate');
         }
+    }
+
+    /**
+     * Setup the test environment.
+     *
+     * @return void
+     */
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        Schema::create('users', function ($table) {
+            $table->id();
+            $table->string('email')->unique();
+            $table->string('password');
+            $table->timestamps();
+        });
+    }
+}
+
+class User extends Model implements Billable
+{
+    use BillableTrait,
+        HasFactory;
+
+    protected static function newFactory()
+    {
+        return UserFactory::new();
+    }
+}
+
+class UserFactory extends Factory
+{
+    protected $model = User::class;
+
+    public function definition()
+    {
+        return [
+            'email' => $this->faker->email(),
+            'password' => $this->faker->password(),
+        ];
     }
 }
